+++ Dockerfile
@@ -2,4 +2,5 @@ FROM python:3.9
 COPY . .
 WORKDIR .
 RUN pip install -r requirements.txt
-CMD uvicorn --reload
+
+CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "80"]
