+++ docker-compose.yml
@@ -1,22 +1,10 @@
 version: "3.9"
 services:
-  db:
-    image: postgres:latest
-    ports:
-      - "${DB_PORT}:5432"
-    environment:
-      - POSTGRES_USER=${DB_USERNAME}
-      - POSTGRES_PASSWORD=${DB_PASSWORD}
-      - POSTGRES_DB=${DB_NAME}
-      - POSTGRES_HOST=${DB_HOST}
   web:
     build: .
-    command: python main.py
     volumes:
       - .:/app
     expose:  # new
       - 8000
-    depends_on:
-      - db
     environment:
       - DB_URL=postgresql://admin:admin@db:5432/api
